Extracts and documents msaValidator.hpp
